Ternary for session page working. Listing only 5 recent sessions on UserHome
