+++ client/src/App.js
@@ -5,6 +5,7 @@ import SignUp from './pages/SignUp';
 import './App.css';
 import Nav from './component/Nav'
 import UserHome from './pages/UserHome'
+import History from './pages/History'
 
 export default function App () {
 
@@ -17,6 +18,23 @@ export default function App () {
 		else if (!userAuth) setUserAuth(true)
 	}
 
+	const optionArray = [{ session: "Running" }, { session: "Studying" }, { session: "Walking" }, { session: "Gaming" }];
+
+	const historyArray = [
+		{ session: "Running", time: "44:15" },
+		{ session: "Walking", time: "1:15:15" },
+		{ session: "Studying", time: "15:10" },
+		{ session: "Reading", time: "45:00" },
+		{ session: "Reading", time: "45:00" },
+		{ session: "Reading", time: "45:00" },
+		{ session: "Reading", time: "45:00" },
+		{ session: "Reading", time: "45:00" },
+		{ session: "Reading", time: "45:00" },
+		{ session: "Reading", time: "45:00" },
+		{ session: "Reading", time: "45:00" },
+
+	];
+
 	return (
 		<div className='App'>
 			{/* <button onClick={authClick}>UserAuth test</button> */}
@@ -24,13 +42,18 @@ export default function App () {
 				<>
 					<Nav authClick={authClick} />
 					<main>
-						<Route exact path="/" component={UserHome} />
+						<Route exact path="/" component={(props) => <UserHome optionArray={optionArray} historyArray={historyArray} />} />
 						<Route exact path='/SignUp' component={SignUp} />
+						<Route exact path='/history' component={(props) => <History historyArray={historyArray} />} />
 						<Route exact path='/About' component={About} />
 					</main>
 				</>
-				: <SignUp authClick={authClick} />}
-
+				: <SignUp authClick={authClick} />
+				// <main>
+				// 	{/* <Route exact path="/signup" component={SignUp} />
+				// 	<Route exact path='/' component={LogIn} /> */}
+				// </main>
+			}
 		</div>
 	);
 }
